+++ packages/jest-runner/src/index.ts
@@ -105,6 +105,7 @@ export default class TestRunner extends EmittingTestRunner {
     }
 
     const worker = new Worker(require.resolve('./testWorker'), {
+      enableWorkerThreads: this._globalConfig.workerThreads,
       exposedMethods: ['worker'],
       forkOptions: {serialization: 'json', stdio: 'pipe'},
       // The workerIdleMemoryLimit should've been converted to a number during