+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## main
 
 ### Features
+
 - `[jest-reporters]` Add `summaryThreshold` option to summary reporter to allow overriding the internal threshold that is used to print the summary of all failed tests when the number of test suites surpasses it ([#13895](https://github.com/facebook/jest/pull/13895))
 
 ### Fixes