+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ### Fixes
 
+- `[jest-circus]` Send test case results for `todo` tests ([#13915](https://github.com/facebook/jest/pull/13915))
+
 ### Chore & Maintenance
 
 ### Performance